fix layer (re)loading issue when switching styles
